dapp-update: reorganized erc constant folders
